refactor(teams): split runner.rs into focused submodules [OPE-366]
Decompose the monolithic runner.rs (1027 LOC) into a runner/ directory
with five focused modules: types (data types and provider resolution),
construction (factory methods), execution (all run variants and provider
activation), output (response parsing), and tests.
